+++ packages/react/src/components/SelectInput/index.tsx
@@ -1,25 +1,14 @@
-import { ComponentProps } from 'react'
+import React, { ComponentProps } from 'react'
 import { Select, SelectInputContainer } from './styles'
 
 export interface SelectInputProps extends ComponentProps<typeof Select> {
-  options: {
-    id: string
-    text: string
-  }[]
+  children: React.ReactNode
 }
 
-export function SelectInput({ options, ...props }: SelectInputProps) {
+export function SelectInput({ children, ...props }: SelectInputProps) {
   return (
     <SelectInputContainer>
-      <Select {...props}>
-        {options?.map((item) => {
-          return (
-            <option value={item.id} key={item.id}>
-              {item.text}
-            </option>
-          )
-        })}
-      </Select>
+      <Select {...props}>{children}</Select>
     </SelectInputContainer>
   )
 }